:recycle: refactor: separe authOptions
